LayoutTests/fast/css/pseudo-class-internal-expected.txt

This tests that -internal- pseudo classes are not exposed

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS target.matches(":-internal-direct-focus") threw exception SyntaxError: The string did not match the expected pattern..
PASS successfullyParsed is true

TEST COMPLETE
